Render error when user is trying to set overridden value over API
