+++ apps/fz_http/lib/fz_http/config/resolver.ex
@@ -11,11 +11,69 @@ defmodule FzHttp.Config.Resolver do
         ) ::
           {:ok, {source :: source(), value :: term()}} | :error
   def resolve(key, env_configurations, db_configurations, opts) do
-    with :error <- resolve_env_value(env_configurations, key, opts),
+    with :error <- resolve_process_env_value(key),
+         :error <- resolve_env_value(env_configurations, key, opts),
          :error <- resolve_db_value(db_configurations, key),
          :error <- resolve_default_value(opts) do
       :error
     end
+  end
+
+  defp resolve_process_env_value(key) do
+    pdict_key = {__MODULE__, key}
+
+    case fetch_process_env(pdict_key) do
+      {:ok, {:env, value}} ->
+        {:ok, {{:env, env_key(key)}, value}}
+
+      :error ->
+        :error
+    end
+  end
+
+  if Mix.env() == :test do
+    def fetch_process_env(pdict_key) do
+      with :error <- fetch_process_value(pdict_key),
+           :error <- fetch_process_value(get_last_pid_from_pdict_list(:"$ancestors"), pdict_key),
+           :error <- fetch_process_value(get_last_pid_from_pdict_list(:"$callers"), pdict_key) do
+        :error
+      end
+    end
+
+    defp fetch_process_value(key) do
+      case Process.get(key) do
+        nil -> :error
+        value -> {:ok, value}
+      end
+    end
+
+    defp fetch_process_value(nil, _key) do
+      :error
+    end
+
+    defp fetch_process_value(atom, key) when is_atom(atom) do
+      atom
+      |> Process.whereis()
+      |> fetch_process_value(key)
+    end
+
+    defp fetch_process_value(pid, key) do
+      with {:dictionary, pdict} <- :erlang.process_info(pid, :dictionary),
+           {^key, value} <- List.keyfind(pdict, key, 0) do
+        value
+      else
+        _other ->
+          :error
+      end
+    end
+
+    defp get_last_pid_from_pdict_list(stack) do
+      if values = Process.get(stack) do
+        List.last(values)
+      end
+    end
+  else
+    def fetch_process_env(app, key), do: :error
   end
 
   defp resolve_env_value(env_configurations, key, opts) do
